Adding L2CCC object to differentiate various network types
Each network type has its own configuration that must be
maintained. When diffing all network types are checked at the same
time; This has led to very complicated diffing methods. Ideally we
will generate a diff on an XML object in the future; This is the main
purpose of this commit, which attempts to use an L2CCC object to
achive this purpose.