Adding `shallow` arg to `to_hash`
Added `shallow` as an arg that could be passed to the `to_hash`
method. The idea was that if provided, the resulting hash would only
include its immediate object vars.

It's not clear to me if this will actually work though; The same issue
(excessive processing) needs to be addressed in the loading from db
phase.

It almost seems that it would be easier to do something like the
following where objects are built up as needed. It's messier but gives
the dev more flexability in crafting their responses.

```
my $user = OESS::User->get(db => $db, username => 'remote_user');

$user->workgroups(OESS::Workgroup->get(db => $db, user_id => $user->id));

foreach my $wg (@{$user->workgroups}) {
  $wg->interfaces(OESS::Interface->get(db => $db, workgroup_id => $wg->id));
}
```