Adding load_* methods
Another idea I had was to require explicit loading of child
resources. This seems to work with the general way we've done things,
but requires a lot of extra work; The issue is that the changes are
required across the entire object model and will affect remove child
resources that are currently available. This is most likely to affect
frontend rendering.